LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-justify-baseline-005-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
FAIL #target > div 2 assert_equals:
<div style="grid-row: 2; grid-column: span 3; justify-self: baseline;" data-offset-x="100">line1<br>line2</div>
offsetLeft expected 100 but got 0
PASS #target > div 3
FAIL #target > div 4 assert_equals:
<div style="grid-row: 4; grid-column: span 3; justify-self: last baseline;" data-offset-x="10">line1<br>line2</div>
offsetLeft expected 10 but got 0
